Moved Models and Public classes to Fido2NetLib.Models project
