+++ Fido2NetLib.Models/AssertionOptions.cs
@@ -46,7 +46,7 @@ namespace Fido2NetLib
         [JsonProperty("extensions", NullValueHandling = NullValueHandling.Ignore)]
         public AuthenticationExtensionsClientInputs Extensions { get; set; }
 
-        internal static AssertionOptions Create(Fido2.Configuration config, byte[] challenge, IEnumerable<PublicKeyCredentialDescriptor> allowedCredentials, UserVerificationRequirement? userVerification, AuthenticationExtensionsClientInputs extensions)
+        public static AssertionOptions Create(Configuration config, byte[] challenge, IEnumerable<PublicKeyCredentialDescriptor> allowedCredentials, UserVerificationRequirement? userVerification, AuthenticationExtensionsClientInputs extensions)
         {
             return new AssertionOptions()
             {
