+++ Fido2NetLib.Models/AuthenticatorAssertionRawResponse.cs
