+++ Fido2NetLib.Models/AuthenticatorAttestationRawResponse.cs
