+++ Fido2NetLib.Models/Fido2ResponseBase.cs
