+++ Fido2NetLib.Models/Fido2VerificationException.cs
