+++ Fido2NetLib.Models/Objects/AttestationVerificationSuccess.cs
