+++ Fido2NetLib.Models/Objects/AuthenticatorAttachment.cs
