+++ Fido2NetLib.Models/Objects/AuthenticatorBiometricPerfBounds.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace Fido2NetLib.Objects
+{
+    public class AuthenticatorBiometricPerfBounds
+    {
+        [JsonProperty("FAR", NullValueHandling = NullValueHandling.Ignore)]
+        public float FAR { get; set; }
+        [JsonProperty("FRR", NullValueHandling = NullValueHandling.Ignore)]
+        public float FRR { get; set; }
+    }
+}
+
