+++ Fido2NetLib.Models/Objects/AuthenticatorTransport.cs
