+++ Fido2NetLib.Models/Objects/TxAuthGenericArg.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace Fido2NetLib.Objects
+{
+    public class TxAuthGenericArg
+    {
+        [JsonProperty("contentType", NullValueHandling = NullValueHandling.Ignore)]
+        public string ContentType { get; set; }
+        [JsonProperty("content", NullValueHandling = NullValueHandling.Ignore)]
+        public byte[] Content { get; set; }
+    }
+}
+
